// src/core/store/modules/deal/utils/log-activities.enum.ts
export enum TYPE_EVENT_LOG {
  CREATION = 'CREATION',
  CHANGE_STATUS = 'CHANGE_STATUS',
  ASSIGNMENT_ASSIGNEE = 'ASSIGNMENT_ASSIGNEE',
  REMOVAL_ASSIGNEE = 'REMOVAL_ASSIGNEE',
  ADDITION_COMMENT = 'ADDITION_COMMENT',
  SCHEDULING_VISIT = 'SCHEDULING_VISIT',
  CONFIRMATION_VISIT = 'CONFIRMATION_VISIT',
  COMPLETION_VISIT = 'COMPLETION_VISIT',
  REMOVAL_VISIT = 'REMOVAL_VISIT',
  CREATION_TASK = 'CREATION_TASK',
  COMPLETION_TASK = 'COMPLETION_TASK',
  REMOVAL_TASK = 'REMOVAL_TASK',
  CHANGE_TYPE_DEAL = 'CHANGE_TYPE_DEAL',
  CHANGE_DATA_CUSTOMER = 'CHANGE_DATA_CUSTOMER',
  ARCHIVING = 'ARCHIVING',
  UNARCHIVING = 'UNARCHIVING',
  CHANGE_TITLE = 'CHANGE_TITLE',
  CHANGE_NOTE = 'CHANGE_NOTE',
  CHANGE_TEMPERATURE = 'CHANGE_TEMPERATURE',
  ADDITION_ATTACHMENT = 'ADDITION_ATTACHMENT',
  REMOVAL_ATTACHMENT = 'REMOVAL_ATTACHMENT',
  SHARE = 'SHARE',
  SUSPENSION_CREATED = 'SUSPENSION_CREATED',
  SUSPENSION_UPDATED = 'SUSPENSION_UPDATED',
  SUSPENSION_REMOVED = 'SUSPENSION_REMOVED',
  DISTRIBUTION_CONFIGURED = 'DISTRIBUTION_CONFIGURED',
  DEALS_REDISTRIBUTED = 'DEALS_REDISTRIBUTED',
  MONITORING_EXECUTED = 'MONITORING_EXECUTED',
  CHANGE_NAME_TASK = 'CHANGE_NAME_TASK',
  CHANGE_NOTES_TASK = 'CHANGE_NOTES_TASK',
  CHANGE_ASSIGNEE_TASK = 'CHANGE_ASSIGNEE_TASK',
  CHANGE_HOUR_START_TASK = 'CHANGE_HOUR_START_TASK',
  CHANGE_HOUR_END_TASK = 'CHANGE_HOUR_END_TASK',
  CHANGE_DATA_TASK = 'CHANGE_DATA_TASK',
}